requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
transformers
torch
pydantic
numpy
slowapi
python-dotenv
google-generativeai
pandas       
openpyxl
python-multipart